LayoutTests/fast/forms/autofocus-focus-only-once-expected.txt


This form control should have a green background and active state:


PASS document.activeElement is document.getElementById("input2")
PASS successfullyParsed is true

TEST COMPLETE
